Rename to `SqlSelectTextNode`.
