+++ metricflow/sql/sql_plan.py
@@ -12,11 +12,12 @@ from metricflow_semantics.collection_helpers.merger import Mergeable
 from metricflow_semantics.dag.id_prefix import IdPrefix, StaticIdPrefix
 from metricflow_semantics.dag.mf_dag import DagId, DagNode, MetricFlowDag
 from metricflow_semantics.sql.sql_exprs import SqlColumnReferenceExpression, SqlExpressionNode
-from metricflow_semantics.sql.sql_table import SqlTable
 from metricflow_semantics.visitor import VisitorOutputT
 from typing_extensions import Self, override
 
+from metricflow.sql.sql_ctas_node import SqlCreateTableAsNode
 from metricflow.sql.sql_select_node import SqlSelectStatementNode
+from metricflow.sql.sql_select_text_node import SqlSelectTextNode
 from metricflow.sql.sql_table_node import SqlTableNode
 
 logger = logging.getLogger(__name__)
@@ -89,7 +90,7 @@ class SqlPlanNodeVisitor(Generic[VisitorOutputT], ABC):
         raise NotImplementedError
 
     @abstractmethod
-    def visit_query_from_clause_node(self, node: SqlSelectQueryFromClauseNode) -> VisitorOutputT:  # noqa: D102
+    def visit_query_from_clause_node(self, node: SqlSelectTextNode) -> VisitorOutputT:  # noqa: D102
         raise NotImplementedError
 
     @abstractmethod
@@ -125,110 +126,6 @@ class SqlSelectColumn:
         return SqlColumnReferenceExpression.from_column_reference(
             column_name=self.column_alias, table_alias=source_table_alias
         )
-
-
-@dataclass(frozen=True, eq=False)
-class SqlSelectQueryFromClauseNode(SqlPlanNode):
-    """An SQL select query that can go in the FROM clause.
-
-    Attributes:
-        select_query: The SQL select query to include in the FROM clause.
-    """
-
-    select_query: str
-
-    @staticmethod
-    def create(select_query: str) -> SqlSelectQueryFromClauseNode:  # noqa: D102
-        return SqlSelectQueryFromClauseNode(
-            parent_nodes=(),
-            select_query=select_query,
-        )
-
-    @classmethod
-    def id_prefix(cls) -> IdPrefix:  # noqa: D102
-        return StaticIdPrefix.SQL_PLAN_QUERY_FROM_CLAUSE_ID_PREFIX
-
-    @property
-    def description(self) -> str:  # noqa: D102
-        return "Read From a Select Query"
-
-    def accept(self, visitor: SqlPlanNodeVisitor[VisitorOutputT]) -> VisitorOutputT:  # noqa: D102
-        return visitor.visit_query_from_clause_node(self)
-
-    @property
-    def as_select_node(self) -> Optional[SqlSelectStatementNode]:  # noqa: D102
-        return None
-
-    @override
-    def nearest_select_columns(self, cte_source_mapping: SqlCteAliasMapping) -> Optional[Sequence[SqlSelectColumn]]:
-        return None
-
-    @property
-    @override
-    def as_sql_table_node(self) -> Optional[SqlTableNode]:
-        return None
-
-    @override
-    def copy(self) -> SqlSelectQueryFromClauseNode:
-        return SqlSelectQueryFromClauseNode(parent_nodes=self.parent_nodes, select_query=self.select_query)
-
-
-@dataclass(frozen=True, eq=False)
-class SqlCreateTableAsNode(SqlPlanNode):
-    """An SQL node representing a CREATE TABLE AS statement.
-
-    Attributes:
-        sql_table: The SQL table to create.
-    """
-
-    sql_table: SqlTable
-
-    def __post_init__(self) -> None:  # noqa: D105
-        super().__post_init__()
-        assert len(self.parent_nodes) == 1
-
-    @staticmethod
-    def create(sql_table: SqlTable, parent_node: SqlPlanNode) -> SqlCreateTableAsNode:  # noqa: D102
-        return SqlCreateTableAsNode(
-            parent_nodes=(parent_node,),
-            sql_table=sql_table,
-        )
-
-    @override
-    def accept(self, visitor: SqlPlanNodeVisitor[VisitorOutputT]) -> VisitorOutputT:
-        return visitor.visit_create_table_as_node(self)
-
-    @property
-    @override
-    def as_select_node(self) -> Optional[SqlSelectStatementNode]:
-        return None
-
-    @property
-    @override
-    def as_sql_table_node(self) -> Optional[SqlTableNode]:
-        return None
-
-    @property
-    @override
-    def description(self) -> str:
-        return f"Create table {repr(self.sql_table.sql)}"
-
-    @property
-    def parent_node(self) -> SqlPlanNode:  # noqa: D102
-        return self.parent_nodes[0]
-
-    @classmethod
-    @override
-    def id_prefix(cls) -> IdPrefix:
-        return StaticIdPrefix.SQL_PLAN_CREATE_TABLE_AS_ID_PREFIX
-
-    @override
-    def nearest_select_columns(self, cte_source_mapping: SqlCteAliasMapping) -> Optional[Sequence[SqlSelectColumn]]:
-        return self.parent_node.nearest_select_columns(cte_source_mapping)
-
-    @override
-    def copy(self) -> SqlCreateTableAsNode:
-        return SqlCreateTableAsNode(parent_nodes=self.parent_nodes, sql_table=self.sql_table)
 
 
 class SqlPlan(MetricFlowDag[SqlPlanNode]):
